requirements.txt
py-cord==2.4.1
orjson
aiodns
brotlipy
PyNaCl
python-dotenv
overfast-py